Allans changes, need to make it work

+++ SqliteGenerationXamarin/SqliteGenerationXamarin/Services/SQLiteFactory.cs
@@ -8,9 +8,20 @@ using HttpTracer;
 using SqliteGeneration.Core;
 using System.Collections.Generic;
 using SQLite;
+using System.Reactive.Linq;
+using System.ComponentModel;
 
 namespace SqliteGenerationXamarin.Services
 {
+    public class Download
+    {
+        public decimal Progress { get; internal set; }
+        public TimeSpan TimeLeft { get; internal set; }
+        public double DownloadSpeedBytes { get; internal set; }
+        public long FileSize { get; internal set; }
+        public long BytesDownloads { get; internal set; }
+    }
+
     public class SQLiteFactory : ISQLiteFactory
     {
         private Action<string> _messageUpdateAction;
@@ -26,6 +37,48 @@ namespace SqliteGenerationXamarin.Services
             _databasePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Personal), "Todo.db");
             _httpTransferTask = httpTransferTask;
         }
+
+
+        public IObservable<Download> DownloadSomeShit() => Observable.Create<Download>(async ob =>
+        {
+
+            DeleteSqliteFile();
+            var url = await this.RequestSqliteGenerationAndDownloadUrl();
+            var task = _httpTransferTask.Download(url.Replace("\"", ""));
+            var progress = new Download();
+
+            var handler = new PropertyChangedEventHandler((sender, args) =>
+            {
+                switch (task.Status)
+                {
+                    case Plugin.HttpTransferTasks.TaskStatus.Error:
+                        ob.OnError(task.Exception);
+                        break;
+
+                    case Plugin.HttpTransferTasks.TaskStatus.Completed:
+                        ob.OnCompleted();
+                        break;
+
+                    default:
+                        progress.FileSize = task.FileSize;
+                        progress.DownloadSpeedBytes = task.BytesPerSecond;
+                        progress.Progress = task.PercentComplete;
+                        progress.TimeLeft = task.EstimatedCompletionTime;
+                        progress.BytesDownloads = task.BytesTransferred;
+
+                        ob.OnNext(progress); 
+                        break;
+                }
+            });
+
+            task.PropertyChanged += handler;
+
+            return () => {
+                task.PropertyChanged -= handler;
+            }; // double fuck off
+        })
+        .Sample(TimeSpan.FromSeconds(1));
+
 
         public async Task DownloadSqlite(Action<string> messageUpdateAction)
         {
@@ -44,18 +97,19 @@ namespace SqliteGenerationXamarin.Services
 
         private async Task<string> RequestSqliteGenerationAndDownloadUrl()
         {
-            _messageUpdateAction("Generating DB on the cloud...");
+            //HttpResponseMessage result;
+            //using (HttpClient client = new HttpClient(new HttpTracerHandler()))
+            //{
+            //    result = await client.GetAsync("http://10.211.55.3:49787/api/data/generateSqlite");
+            //}
 
-            HttpResponseMessage result;
-            using (HttpClient client = new HttpClient(new HttpTracerHandler()))
-            {
-                result = await client.GetAsync("http://10.211.55.3:49787/api/data/generateSqlite");
-            }
+            return "https://bsistoragedemocauser.blob.core.windows.net/sqlitefile/fuckoffshahab.db";
+            //result.EnsureSuccessStatusCode();
+            //return await result.Content.ReadAsStringAsync();
+            //if (result.IsSuccessStatusCode)
+            //    return await result.Content.ReadAsStringAsync();
 
-            if (result.IsSuccessStatusCode)
-                return await result.Content.ReadAsStringAsync();
-
-            throw new Exception("Could not Generate sqlite on the cloud.");
+            //throw new Exception("Could not Generate sqlite on the cloud.");
         }
 
         private void DownloadSqliteFile(string downloadUrl)
@@ -74,7 +128,8 @@ namespace SqliteGenerationXamarin.Services
 
         public void DeleteSqliteFile()
         {
-            File.Delete(_databasePath);
+            if (File.Exists(_databasePath))
+                File.Delete(_databasePath);
         }
 
         public List<TodoItem> FetchTodoData()
